Adding child panels to existing comic
This adds logic to find the parent panel and add the new child panel
to it when a user adds panels to an existing comic.

Threw in janky unit test to make sure the recursion works as expected,
will be adding more testing later.

synced up common folder in other lambdas