Adding logic for presigned url
This url will allow grant temp access to the user to upload their image
to s3.

The pattern will be comics/{comic_id}/{panel_id}.jpg
I'm setting the first panel id to always be the same as the comic
id, since this will make it easier to fetch the first image for the
main page.